+++ api_tests/e2e/rfc003/eth_btc/happy.ts
@@ -5,7 +5,7 @@ import { Actor } from "../../../lib/actor";
 import { Action, SwapRequest, SwapResponse } from "../../../lib/comit";
 import { BN, toBN, toWei } from "web3-utils";
 import { HarnessGlobal } from "../../../lib/util";
-import { ActionTrigger, AfterTest, execute } from "../../test_executor";
+import { ActionTrigger, AfterTest, createTests } from "../../test_creator";
 import chaiHttp = require("chai-http");
 
 const should = chai.should();
@@ -138,7 +138,7 @@ async function test() {
     ];
 
     describe("RFC003: Ether for Bitcoin", () => {
-        execute(alice, bob, actions, initialUrl, listUrl, swapRequest);
+        createTests(alice, bob, actions, initialUrl, listUrl, swapRequest);
     });
     run();
 }
